[Oztechan/CCC#1831] Create :client:repository:background module
