+++ client/repository/background/src/commonMain/kotlin/com/oztechan/ccc/client/repository/background/BackgroundRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.oztechan.ccc.client.repository.background
 
 import co.touchlab.kermit.Logger
+import com.oztechan.ccc.client.core.shared.util.getRateFromCode
 import com.oztechan.ccc.client.datasource.watcher.WatcherDataSource
 import com.oztechan.ccc.client.service.backend.BackendApiService
-import com.oztechan.ccc.client.util.getRateFromCode
 import kotlinx.coroutines.runBlocking
 
 internal class BackgroundRepositoryImpl(
