+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/widget/WidgetViewModel.kt
@@ -1,6 +1,7 @@
 package com.oztechan.ccc.client.viewmodel.widget
 
 import com.oztechan.ccc.client.base.BaseViewModel
+import com.oztechan.ccc.client.core.shared.util.getRateFromCode
 import com.oztechan.ccc.client.core.shared.util.isItOver
 import com.oztechan.ccc.client.core.shared.util.nowAsInstant
 import com.oztechan.ccc.client.core.shared.util.toDateString
@@ -9,7 +10,6 @@ import com.oztechan.ccc.client.service.backend.BackendApiService
 import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
 import com.oztechan.ccc.client.util.getFormatted
-import com.oztechan.ccc.client.util.getRateFromCode
 import com.oztechan.ccc.client.util.toValidList
 
 class WidgetViewModel(
