+++ ios/provider/build.gradle.kts
@@ -78,6 +78,7 @@ kotlin {
 
                 Modules.Client.Repository.apply {
                     implementation(project(adControl))
+                    implementation(project(background))
                 }
 
                 implementation(project(Modules.Submodules.logmob))
